Add files via upload
This code is a simple guessing game where the user tries to guess a randomly generated mystery number between 0 and 100. The game provides feedback on whether the guessed number is too large, too small, or correct. The user has a limited number of attempts (5 in this case) to guess the mystery number.